% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbrvalues.R
\name{nbrs_sum}
\alias{nbrs_sum}
\title{Add sum of neighbors values to SF object}
\usage{
nbrs_sum(sf_object, unique_id_col, variable, date_col)
}
\arguments{
\item{sf_object}{An object of class 'SF'}

\item{unique_id_col}{A string identifying a unique column to serve as an identifier}

\item{variable}{A string identifying a numeric variable}

\item{date_col}{(Optional) A string identifying a variable for date, use if data is over multiple time periods.}
}
\value{
Column string of neighbors
}
\description{
Add sum of neighbors values to SF object
}
\examples{
panel_example_sf <- nbrs_sum(sf_object = panel_example_sf, unique_id_col = "country", variable = "gdp", date_col="year")
panel_example_sf <- nbrs_sum(panel_example_sf, "country", "gdp", "year")
cross_section_example_sf <- nbrs_sum(sf_object = cross_section_example_sf, unique_id_col = "country", variable = "gdp")
cross_section_example_sf <- nbrs_sum(cross_section_example_sf, "country", "gdp")
}
